refactor: extract class `DebuggingPreferencesDataStore.java` from `DebuggingPreferencesFragment.java` and make new package `datastore`
`DebuggingPreferencesDataStore.java` have two class. Extracting `DebuggingPreferencesDataStore.java` help maintain adherence to the Single Responsibility Principle. and make `datastore` package.
